fastapi_jwt_auth/requirements.txt
bcrypt==4.0.0
cryptography==37.0.*
deta==1.1.0
fastapi==0.81.*
passlib==1.7.*
pydantic[email]
pyjwt==2.4.0
pytest==7.1.3
python-dotenv==0.20.0
requests==2.28.1
uvicorn
